tests/wpt/tests/css/css-tables/visibility-collapse-col-004-dynamic.html

<!doctype html>
<meta charset="utf-8">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<link rel="author" title="Joy Yu" href="mailto:joysyu@mit.edu">
<link rel="help" href="https://drafts.csswg.org/css-tables-3/#computing-the-table-width">
<style>
    table {
      border: 5px solid black;
    }
    table span {
        display: inline-block;
        vertical-align: top;
        background: lime;
        margin: 3px;
        height: 100px;
        width: 100px;
    }
</style>
<main>
    <h1>Dynamic Visibility collapse</h1>
    <a href="https://drafts.csswg.org/css-tables-3/#computing-the-table-width">Spec</a>
    <p>
    Setting a column to visibility:collapse changes table width but not height.
    Bottom table is identical to top except left column has been collapsed.
    </p>
    <table id="test">
      <colgroup id="collapse"></colgroup>
      <tbody>
        <tr>
            <td>
                <span>row 1</span>
            </td>
            <td>
                <span></span>
            </td>
        </tr>
        <tr>
            <td>
                <span>row 2</span>
            </td>
            <td>
                <span></span>
            </td>
        </tr>
        <tr>
            <td>
                <span>row 3</span>
            </td>
            <td>
                <span style="height:50px"></span>
            </td>
        </tr>
      </tbody>
    </table>
</main>

<script>
    colgroup = document.getElementById("collapse");
    colgroup.style.visibility = "collapse";

    tests = [
        [
            document.getElementById('test').offsetHeight,
            342,
            "col visibility:collapse doesn't change table height",
        ],
        [
            document.getElementById('test').offsetWidth,
            122,
            "col visibility:collapse changes table width"
        ]];
    for (i = 0; i< tests.length; i++) {
        test(function()
            {
                assert_equals.apply(this, tests[i]);
            },
            tests[i][2]);
    };
</script>
</html>
