LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-col-004-dynamic-expected.txt

Dynamic Visibility collapse

Spec
Setting a column to visibility:collapse changes table width but not height. Bottom table is identical to top except left column has been collapsed.

row 1	
row 2	
row 3	

PASS col visibility:collapse doesn't change table height
FAIL col visibility:collapse changes table width assert_equals: col visibility:collapse changes table width expected 122 but got 232
